[ENH] Split param_est.stationarity._arch into one-estimator-per-file submodules
