+++ sktime/param_est/stationarity/__init__.py
@@ -13,16 +13,14 @@ __all__ = [
     "BreakvarHeteroskedasticityTest",
 ]
 
-from sktime.param_est.stationarity._arch import (
-    StationarityADFArch,
-    StationarityDFGLS,
-    StationarityKPSSArch,
-    StationarityPhillipsPerron,
-    StationarityVarianceRatio,
-    StationarityZivotAndrews,
-)
+from sktime.param_est.stationarity._adf_arch import StationarityADFArch
+from sktime.param_est.stationarity._dfgls import StationarityDFGLS
+from sktime.param_est.stationarity._kpss_arch import StationarityKPSSArch
+from sktime.param_est.stationarity._phillips_perron import StationarityPhillipsPerron
 from sktime.param_est.stationarity._sm_breakvar import BreakvarHeteroskedasticityTest
 from sktime.param_est.stationarity._statsmodels import (
     StationarityADF,
     StationarityKPSS,
 )
+from sktime.param_est.stationarity._variance_ratio import StationarityVarianceRatio
+from sktime.param_est.stationarity._zivot_andrews import StationarityZivotAndrews
